Refactor Videos component into its own folder
