+++ src/videos/components/Videos/Videos.js
@@ -1,7 +1,7 @@
 import React, { Component, Fragment } from 'react'
 import { Link } from 'react-router-dom'
 
-import { getVideos } from '../api'
+import { getVideos } from '../../api'
 
 import './Videos.scss'
